+++ test/support/mocks/mongo-client-mock.ts
@@ -5,7 +5,8 @@ const mongoClientMock = (sandbox: SinonSandbox) => ({
   find: sandbox.stub(),
   toArray: sandbox.stub(),
   findOne: sandbox.stub(),
-  collection: sandbox.stub()
+  collection: sandbox.stub(),
+  isConnected: sandbox.stub()
 })
 
 export { mongoClientMock }